Sidebar
Added sidebar functionality. This comprised of moving the login and sign-up buttons and their function into the sidebar. The sidebar begins off the screen but there is a button that you can click to make it slide onto the screen. For future development, make the button look nicer and the sidebar more feature-filled, potentially with a logo created for recipify